* develop: (30 commits)
  แก้ไข static
  add comment
  เปลี่ยน label ให้เป็นภาษา EN
  เปลี่ยน file default.
  เปลี่ยนลิ้งค์เรียก requisition
  แก้ไข receive info ได้
  แก้ไขเรื่องการ cutStock ของ requisition
  ไม่ให้เพิ่ม Product เข้าไปได้อีก ถ้าได้เพิ่มเข้าไปรอบนึ่งแล้ว
  default min date สำหรับเลือกวันที่รับใบเบิก
  เปลี่ยน Product List ให้เป็น Datatable
  เมื่อสร้างใบเบิกให้ไปหน้า Add Product Auto
  add receive
  ลบ Code ที่ไม่ได้ใช้
  การเพิ่ม location จะเพิ่ม stock สำหรับ location เป็น default ให้ด้วยที่ตาราง stocks.
  sortable table receive requesition.
  add field group, number on requesition item.
  เปลี่ยนการแสดงผลของหน้า process
  update item for requesition.
  fix url
  export excel หน้า movement.
  ...

+++ app/Location.php
@@ -4,38 +4,72 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 
+/**
+ * App\Location
+ *
+ * @property-read \Illuminate\Database\Eloquent\Collection|\App\ReceiveItem[] $receiveItems
+ * @property-read \Illuminate\Database\Eloquent\Collection|\App\Stock[] $stocks
+ * @property integer $id
+ * @property string $name
+ * @property \Carbon\Carbon $created_at
+ * @property \Carbon\Carbon $updated_at
+ */
 class Location extends Model
 {
     protected $fillable = [
-    	'name',
+        'name',
     ];
+
+    public static function boot()
+    {
+        parent::boot();
+
+        static::created(function ($model) {
+            $products = Product::all();
+            foreach ($products as $product) {
+                $model->stocks()->create([
+                    'product_id' => $product->id,
+                    'qty'        => 0,
+                ]);
+            }
+        });
+    }
 
     public function receiveItems()
     {
-    	return $this->hasMany(ReceiveItem::class, 'location_id', 'id');
+        return $this->hasMany(ReceiveItem::class, 'location_id', 'id');
+    }
+
+    public function stocks()
+    {
+        return $this->hasMany(Stock::class);
     }
 
     public static function deleteByCondition($id)
     {
-    	$query = self::with([
-    			'receiveItems',
-    		])
-    		->where('id', $id)
-    		->first();
+        $query = self::with([
+            'receiveItems',
+            'stocks' => function ($query) {
+                $query->where('qty', '>', 0);
+            },
+        ])
+            ->where('id', $id)
+            ->first();
 
-    	if($query->receiveItems()->count())
-			return [
-				'status' => false,
-				'title' => trans('location.label.name'),
-				'message' => trans('location.message_alert.delete_unsuccess'),
-			];
+        if ($query->receiveItems->count() || $query->stocks->count()) {
+            return [
+                'status'  => false,
+                'title'   => trans('location.label.name'),
+                'message' => trans('location.message_alert.delete_unsuccess'),
+            ];
+        }
 
-		$response = $query->delete();
+        $response = $query->delete();
 
-		return [
-			'status' => $response,
-			'title' => trans('location.label.name'),
-			'message' => trans('location.message_alert.delete_success'),
-		];
+        return [
+            'status'  => $response,
+            'title'   => trans('location.label.name'),
+            'message' => trans('location.message_alert.delete_success'),
+        ];
     }
 }
